requirements.txt
boto3>=1.26.0
fastapi>=0.104.1
locust>=2.24.0
numpy>=1.26.0
pesq>=0.0.4
psutil>=5.9.0
pydub>=0.25.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest>=7.4.0
python-dotenv>=1.0.0
python-magic>=0.4.27
python-multipart
requests>=2.31.0
scipy>=1.11.3
soundfile>=0.12.1
uvicorn>=0.24.0